feat(macro): add babel-plugin-macros compatible macro
