+++ prettier.config.js
@@ -1,0 +1,2 @@
+// here only so the editor can find our config
+module.exports = require('kcd-scripts/dist/config/prettierrc')
