+++ src/index.js
@@ -1,15 +1,9 @@
-const nodePath = require('path')
-// const nodePath = {} // handy when you wanna copy this into astexplorer.net
+import handleGlamorousReferences from './handle-glamorous-references'
+import looksLike from './looks-like'
 
 export default function(babel) {
-  const {types: t, template} = babel
-  const identifiers = new Set()
-  const buildBuiltInWithConfig = template(`
-    GLAMOROUS.BUILT_IN.withConfig({displayName: DISPLAY_NAME})
-  `)
-  const buildCustomWithConfig = template(`
-    GLAMOROUS(ARGUMENTS).withConfig({displayName: DISPLAY_NAME})
-  `)
+  const {types: t} = babel
+  const references = new Set()
   return {
     name: 'babel-plugin-glamorous-displayName',
     visitor: {
@@ -24,7 +18,7 @@ export default function(babel) {
         const {node: {local: {name}}} = defaultSpecifierPath
         const {referencePaths} = path.scope.getBinding(name)
         referencePaths.forEach(reference => {
-          identifiers.add(reference)
+          references.add(reference)
         })
       },
       VariableDeclarator(path) {
@@ -40,116 +34,17 @@ export default function(babel) {
         if (binding) {
           const {referencePaths} = binding
           referencePaths.forEach(reference => {
-            identifiers.add(reference)
+            references.add(reference)
           })
         }
       },
       Program: {
         exit(path, {file}) {
-          Array.from(identifiers).forEach(identifier => {
-            const displayName = getDisplayName(identifier, file)
-
-            handleBuiltIns(identifier, displayName)
-            handleCustomComponent(identifier, displayName)
-          })
+          handleGlamorousReferences(Array.from(references), file, babel)
         },
       },
     },
   }
-
-  function handleBuiltIns(path, displayName) {
-    const isBuiltIn = looksLike(path, {
-      parentPath: {
-        type: 'MemberExpression',
-        node: {
-          property: {
-            type: 'Identifier',
-          },
-        },
-        parent: {
-          type: 'CallExpression',
-        },
-      },
-    })
-    if (!isBuiltIn) {
-      return
-    }
-    path.parentPath.replaceWith(
-      buildBuiltInWithConfig({
-        GLAMOROUS: path.node,
-        BUILT_IN: path.parent.property,
-        DISPLAY_NAME: t.stringLiteral(displayName),
-      }),
-    )
-  }
-
-  function handleCustomComponent(path, displayName) {
-    const isCustom = looksLike(path, {
-      parent: {
-        type: 'CallExpression',
-      },
-    })
-    if (!isCustom) {
-      return
-    }
-    path.parentPath.replaceWith(
-      buildCustomWithConfig({
-        GLAMOROUS: path.node,
-        ARGUMENTS: path.parent.arguments,
-        DISPLAY_NAME: t.stringLiteral(displayName),
-      }),
-    )
-  }
-
-  // credit: https://github.com/styled-components/babel-plugin-styled-components/blob/37a13e9c21c52148ce6e403100df54c0b1561a88/src/visitors/displayNameAndId.js
-  function getDisplayName(path, file) {
-    const componentName = getName(path)
-    const filename = getFileName(file)
-    if (filename) {
-      if (filename === componentName) {
-        return componentName
-      }
-      return componentName ? `${filename}__${componentName}` : filename
-    } else {
-      return componentName
-    }
-  }
-
-  // credit: https://github.com/styled-components/babel-plugin-styled-components/blob/37a13e9c21c52148ce6e403100df54c0b1561a88/src/utils/getName.js
-  function getName(path) {
-    let namedNode
-
-    path.find(parentPath => {
-      if (parentPath.isObjectProperty()) {
-        // const X = { Y: glamorous }
-        namedNode = parentPath.node.key
-      } else if (parentPath.isVariableDeclarator()) {
-        // let X; X = glamorous
-        namedNode = parentPath.node.id
-      } else if (parentPath.isStatement()) {
-        // we've hit a statement, we should stop crawling up
-        return true
-      }
-
-      // we've got an displayName (if we need it) no need to continue
-      if (namedNode) {
-        return true
-      }
-      return false
-    })
-
-    // identifiers are the only thing we can reliably get a name from
-    return t.isIdentifier(namedNode) ? namedNode.name : undefined
-  }
-}
-
-function getFileName(file) {
-  if (!file || file.opts.filename === 'unknown') {
-    return ''
-  }
-  return file.opts.basename === 'index'
-    ? nodePath.basename(nodePath.dirname(file.opts.filename))
-    : file.opts.basename
 }
 
 function isRequireCall(callExpression) {
@@ -161,24 +56,4 @@ function isRequireCall(callExpression) {
     arguments: args =>
       args && args.length === 1 && args[0].value === 'glamorous',
   })
-}
-
-function looksLike(a, b) {
-  return (
-    a &&
-    b &&
-    Object.keys(b).every(bKey => {
-      const bVal = b[bKey]
-      const aVal = a[bKey]
-      if (typeof bVal === 'function') {
-        return bVal(aVal)
-      }
-      return isPrimitive(bVal) ? bVal === aVal : looksLike(aVal, bVal)
-    })
-  )
-}
-
-function isPrimitive(val) {
-  // eslint-disable-next-line
-  return val == null || /^[sbn]/.test(typeof val)
 }
